+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/Accessory/AccessoryFileReader.cs
@@ -1,3 +1,5 @@
+using System;
+using System.IO;
 using UniGLTF;
 using UnityEngine;
 using UnityEngine.Rendering;
@@ -24,16 +26,26 @@ namespace Baku.VMagicMirror
         // NOTE: ぜんぶ同じフォルダに入ってる事は保証されてない事に注意。
         // gltfや、(今は無いけど想定される例として)連番画像とかはフォルダを区切った中に入る。
 
-        public static AccessoryFileContext<Texture2D> LoadPngImage(byte[] bytes)
+        public static AccessoryFileContext<Texture2D> LoadPngImage(AccessoryFile file, byte[] bytes)
         {
+            if (bytes == null || bytes.Length == 0)
+            {
+                return new AccessoryFileContext<Texture2D>(null, new ImageAccessoryActions(file, null));
+            }
+
             var tex = new Texture2D(8, 8, TextureFormat.RGBA32, false);
             tex.LoadImage(bytes);
             tex.Apply();
-            return new AccessoryFileContext<Texture2D>(tex, new ImageAccessoryActions(tex));
+            return new AccessoryFileContext<Texture2D>(tex, new ImageAccessoryActions(file, tex));
         }
 
         public static AccessoryFileContext<GameObject> LoadGltf(string path, byte[] bytes)
         {
+            if (!File.Exists(path) || bytes == null || bytes.Length == 0)
+            {
+                return new AccessoryFileContext<GameObject>(new GameObject("(empty)"), new EmptyFileActions());
+            }
+
             var parser = new AutoGltfFileParser(path);
             using var data = parser.Parse();
             return LoadGlbOrGltf(data);
@@ -41,6 +53,11 @@ namespace Baku.VMagicMirror
 
         public static AccessoryFileContext<GameObject> LoadGlb(string path, byte[] bytes)
         {
+            if (!File.Exists(path) || bytes == null || bytes.Length == 0)
+            {
+                return new AccessoryFileContext<GameObject>(new GameObject("(empty)"), new EmptyFileActions());
+            }
+
             var parser = new GlbLowLevelParser("", bytes);
             using var data = parser.Parse();
             return LoadGlbOrGltf(data);
@@ -48,6 +65,10 @@ namespace Baku.VMagicMirror
 
         public static AccessoryFileContext<AnimatableImage> LoadNumberedPngImage(byte[][] binaries)
         {
+            if (binaries == null)
+            {
+                binaries = Array.Empty<byte[]>();
+            }
             var res = new AnimatableImage(binaries);
             //他と違い、AnimatableImage自体がFileActionを実装済み
             return new AccessoryFileContext<AnimatableImage>(res, res);