+++ WPF/VMagicMirrorConfig/Model/SettingModel/AccessorySettingModel.cs
@@ -103,6 +103,8 @@ namespace Baku.VMagicMirrorConfig
                 item.Rotation = Vector3.Zero();
                 item.Scale = Vector3.One();
                 item.Name = Path.GetFileNameWithoutExtension(item.FileId);
+                item.ResolutionLimit = AccessoryImageResolutionLimit.None;
+                item.UseAsBlinkEffect = false;
                 ItemUpdated?.Invoke(item);
             }
             SerializedSetting = JsonConvert.SerializeObject(Items);
@@ -157,6 +159,7 @@ namespace Baku.VMagicMirrorConfig
             target.Position = item.Position;
             target.Rotation = item.Rotation;
             target.Scale = item.Scale;
+            target.ResolutionLimit = item.ResolutionLimit;
             ItemUpdated?.Invoke(target);
         }